// src/app/api/models/dto/CreateAtividadeReuniaoEventoDto.ts
export class CreateAtividadeReuniaoEventoDto
{
  tipoNaturezaOcorrencia?: string;
  acaoSubTipoAtividadeReuniaoEvento?: string;
  exemplosOuReferencias?: string;
  representantesNisac?: string;
  representantesExternos?: string;
  entidadePromotoraEvento?: string;
  responsavelEntidadePromotoraEvento?: string;
  local?: string;
  ambitoOuObservacao?: string;
  numeroActa?: string;
}